fix and test

+++ src/net/bigtangle/core/Monetary.ts
@@ -1,5 +1,3 @@
-import { BigInteger  } from 'big-integer';
-
 /**
  * Classes implementing this interface represent a monetary value, such as a Bitcoin or fiat amount.
  */
@@ -7,7 +5,7 @@ export interface Monetary {
     /**
      * Returns the number of "smallest units" of this monetary value. For Bitcoin, this would be the number of satoshis.
      */
-    getValue(): BigInteger;
+    getValue(): bigint;
 
     signum(): number;
 }
